// packages/aq-fe-framework/src/types/type_action.ts
export type type_action =
    | "default"
    | "create"
    | "update"
    | "delete"
    | "save"
    | "createMultiple"
    | "import"
    | "print"
    | "cancel"
    | "export"
    | "select"
    | "find"
    | "tempDelete"
    | "tempUpdate"
    | "viewFile"
    | "sendMail"
    | "view"
    | "validate"
    | "sync"